Take float value instead of int.

+++ src/main/java/com/nadmm/airports/utils/UiUtils.java
@@ -59,7 +59,7 @@ public class UiUtils {
         }
     }
 
-    public static int convertDpToPx( Context context, int dp ) {
+    public static int convertDpToPx( Context context, float dp ) {
         return (int) TypedValue.applyDimension( TypedValue.COMPLEX_UNIT_DIP,
                 dp, context.getResources().getDisplayMetrics() );
     }
